requirements.txt
aniso8601==9.0.1
APScheduler==3.11.1
blinker==1.9.0
click==8.1.7
dnspython==2.7.0
Flask==3.1.0
Flask-RESTful==0.3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
pymongo==4.10.1
python-dotenv==1.0.1
pytz==2024.2
six==1.16.0
tzlocal==5.3.1
Werkzeug==3.1.3
